Add env vars saying whether to talk to MQ & Kafka
Only publish messages to MQ if the Messaging microservice is deployed.  And only publish messages to Kafka if the Trade History microservice is deployed.  Also, only try calling the return-on-investment REST service in Trade History if that is indeed deployed.  This should clean up a lot of the noise in the Portfolio log when either of these optional pieces aren't selected for deployment.