Add 'quote' filter to password related credentials
When performing "templating" (i.e. `helm template`), we hit the issue
related to null value conversion. This has been confirmed by the Helm
community. The workaround is to make sure that empty string values are
passed to the `quote` filter. This will ensure that we will always get
an empty string rather than a `null` value which is invalid for
Kubernetes.